+++ src/main/java/com/basestudy/rewards/constants/ErrorCode.java
@@ -1,14 +1,16 @@
-package com.basestudy.rewards.contants;
+package com.basestudy.rewards.constants;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public enum ResponseMessage {
-    ACCEPTED(202, "202", "", ""),
-    CREATED(201, "201", "", "등록되었습니다.");
-
+public enum ErrorCode {
+    
+    BAD_REQUEST(400, "400", "", ""),
+    UNAUTHORIZED(401, "401", "", ""),
+    FORBIDDEN(403, "403", "", "");
+    
     private int status;
     private String code;
     private String message;
